+++ src/DirectLink/Gateway.php
@@ -51,24 +51,24 @@ class Pronamic_WP_Pay_Gateways_Ogone_DirectLink_Gateway extends Pronamic_WP_Pay_
 
 		$ogone_data_general
 			->set_psp_id( $this->client->psp_id )
-			->set_order_id( Pronamic_WP_Pay_Gateways_Ogone_Util::get_order_id( $this->config->order_id, $data, $payment ) )
-			->set_order_description( $data->get_description() )
+			->set_order_id( Pronamic_WP_Pay_Gateways_Ogone_Util::get_order_id( $this->config->order_id, $payment ) )
+			->set_order_description( $payment->get_description() )
 			->set_param_plus( 'payment_id=' . $payment->get_id() )
-			->set_currency( $data->get_currency() )
-			->set_amount( $data->get_amount() )
-			->set_language( $data->get_language_and_country() );
+			->set_currency( $payment->get_currency() )
+			->set_amount( $payment->get_amount() )
+			->set_language( $payment->get_locale() );
 
 		// Customer
 		$ogone_data_customer = new Pronamic_WP_Pay_Gateways_Ogone_DataCustomerHelper( $ogone_data );
 
 		$ogone_data_customer
-			->set_name( $data->get_customer_name() )
-			->set_email( $data->get_email() )
-			->set_address( $data->get_address() )
-			->set_zip( $data->get_zip() )
-			->set_town( $data->get_city() )
-			->set_country( $data->get_country() )
-			->set_telephone_number( $data->get_telephone_number() );
+			->set_name( $payment->get_customer_name() )
+			->set_email( $payment->get_email() )
+			->set_address( $payment->get_address() )
+			->set_zip( $payment->get_zip() )
+			->set_town( $payment->get_city() )
+			->set_country( $payment->get_country() )
+			->set_telephone_number( $payment->get_telephone_number() );
 
 		// DirectLink
 		$ogone_data_directlink = new Pronamic_WP_Pay_Gateways_Ogone_DirectLink_DataHelper( $ogone_data );
@@ -78,6 +78,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DirectLink_Gateway extends Pronamic_WP_Pay_
 			->set_password( $this->client->password );
 
 		// Credit card
+		/*
 		$ogone_data_credit_card = new Pronamic_WP_Pay_Gateways_Ogone_DataCreditCardHelper( $ogone_data );
 
 		$credit_card = $data->get_credit_card();
@@ -86,6 +87,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_DirectLink_Gateway extends Pronamic_WP_Pay_
 			->set_number( $credit_card->get_number() )
 			->set_expiration_date( $credit_card->get_expiration_date() )
 			->set_security_code( $credit_card->get_security_code() );
+		*/
 
 		$ogone_data->set_field( 'OPERATION', 'SAL' );