+++ src/OrderStandard/Gateway.php
@@ -35,7 +35,7 @@ class Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Gateway extends Pronamic_WP_P
 
 		$this->client = new Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Client( $this->config->psp_id );
 
-		$this->client->set_payment_server_url( $config->get_payment_server_url() );
+		$this->client->set_payment_server_url( $config->get_form_action_url() );
 		$this->client->set_pass_phrase_in( $config->sha_in_pass_phrase );
 		$this->client->set_pass_phrase_out( $config->sha_out_pass_phrase );
 
@@ -45,6 +45,30 @@ class Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Gateway extends Pronamic_WP_P
 	}
 
 	/////////////////////////////////////////////////
+
+	/**
+	 * Get payment methods
+	 *
+	 * @since unreleased
+	 * @return mixed an array or null
+	 */
+	public function get_payment_methods() {
+		$methods_class = 'Pronamic_WP_Pay_Gateways_Ogone_PaymentMethods';
+
+		if ( class_exists( $methods_class ) ) {
+			$payment_methods = new ReflectionClass( $methods_class );
+
+			$groups = array(
+				array(
+					'options' => $payment_methods->getConstants(),
+				),
+			);
+
+			return $groups;
+		}
+
+		return null;
+	}
 
 	/**
 	 * Get supported payment methods
@@ -64,10 +88,9 @@ class Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Gateway extends Pronamic_WP_P
 	/**
 	 * Start
 	 *
-	 * @param Pronamic_Pay_PaymentDataInterface $data
 	 * @see Pronamic_WP_Pay_Gateway::start()
 	 */
-	public function start( Pronamic_Pay_PaymentDataInterface $data, Pronamic_Pay_Payment $payment, $payment_method = null ) {
+	public function start( Pronamic_Pay_Payment $payment ) {
 		$payment->set_action_url( $this->client->get_payment_server_url() );
 
 		$ogone_data = $this->client->get_data();
@@ -76,28 +99,28 @@ class Pronamic_WP_Pay_Gateways_Ogone_OrderStandard_Gateway extends Pronamic_WP_P
 		$ogone_data_general = new Pronamic_WP_Pay_Gateways_Ogone_DataGeneralHelper( $ogone_data );
 
 		$ogone_data_general
-			->set_order_id( Pronamic_WP_Pay_Gateways_Ogone_Util::get_order_id( $this->config->order_id, $data, $payment ) )
-			->set_order_description( $data->get_description() )
+			->set_order_id( Pronamic_WP_Pay_Gateways_Ogone_Util::get_order_id( $this->config->order_id, $payment ) )
+			->set_order_description( $payment->get_description() )
 			->set_param_plus( 'payment_id=' . $payment->get_id() )
-			->set_currency( $data->get_currency() )
-			->set_amount( $data->get_amount() )
-			->set_language( $data->get_language_and_country() );
+			->set_currency( $payment->get_currency() )
+			->set_amount( $payment->get_amount() )
+			->set_language( $payment->get_locale() );
 
 		// Customer
 		$ogone_data_customer = new Pronamic_WP_Pay_Gateways_Ogone_DataCustomerHelper( $ogone_data );
 
 		$ogone_data_customer
-			->set_name( $data->get_customer_name() )
-			->set_email( $data->get_email() )
-			->set_address( $data->get_address() )
-			->set_zip( $data->get_zip() )
-			->set_town( $data->get_city() )
-			->set_country( $data->get_country() )
-			->set_telephone_number( $data->get_telephone_number() );
+			->set_name( $payment->get_customer_name() )
+			->set_email( $payment->get_email() )
+			->set_address( $payment->get_address() )
+			->set_zip( $payment->get_zip() )
+			->set_town( $payment->get_city() )
+			->set_country( $payment->get_country() )
+			->set_telephone_number( $payment->get_telephone_number() );
 
 		// Payment method
 		// @see https://github.com/wp-pay-gateways/ogone/wiki/Brands
-		switch ( $payment_method ) {
+		switch ( $payment->get_method() ) {
 			case Pronamic_WP_Pay_PaymentMethods::CREDIT_CARD :
 				/*
 				 * Set credit card payment method.